backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
supabase==2.4.6
python-jose[cryptography]==3.3.0
pydantic-settings==2.3.4
python-dotenv==1.0.1
httpx==0.27.0
deep-translator==1.11.4
